# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Ai-lifeplanner API",
        "title": "Fiber API AudioSum",
        "contact": {},
        "version": "1.0"
    },
    "host": "humorous-colt-vital.ngrok-free.app",
    "basePath": "/",
    "paths": {
        "/api/v1/ai_gen/ai_gen/{id}": {
            "get": {
                "description": "get Gennerated plan that generate by ai base on ai prompt that we create from base user input",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai Gen"
                ],
                "summary": "Get AI gennerated plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            }
        },
        "/api/v1/ai_gen/chat/{id}": {
            "get": {
                "description": "Get AI Chat history By userID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai Gen"
                ],
                "summary": "Get AI Chat By userID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            },
            "post": {
                "description": "Sent the new message to AI Chat Assistant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai Gen"
                ],
                "summary": "Sent a new message to AI Chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "AI Chat Data",
                        "name": "bodyHealthBackground",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.AIChatResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete all history AI Chat By userID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai Gen"
                ],
                "summary": "Delete AI Chat By userID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            }
        },
        "/api/v1/ai_gen/create_ai_gen/{id}": {
            "post": {
                "description": "generate a new AI GenPlan base ai prompt that we create base on data that user input",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai Gen"
                ],
                "summary": "Create a new AI GenPlan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            }
        },
        "/api/v1/ai_gen/goal/{id}": {
            "delete": {
                "description": "Delete AI goal that we generate from ai By ID of that goal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai Gen"
                ],
                "summary": "Delete AI GeneratedPlan By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GeneratedPlan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/finance_info": {
            "get": {
                "description": "Get finance records of all user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance"
                ],
                "summary": "Get All user finance records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            }
        },
        "/api/v1/users/finance_info/{id}": {
            "get": {
                "description": "Get finance records by User ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance"
                ],
                "summary": "Get finance records by User ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            },
            "post": {
                "description": "create new finance records by User ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance"
                ],
                "summary": "create new finance records by User ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Finance Data",
                        "name": "bodyFinance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.FinanceRespond"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.AIChatResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "entities.FinanceRespond": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "expenses": {
                    "type": "number"
                },
                "income": {
                    "type": "number"
                },
                "risk_tolerance": {
                    "type": "string"
                },
                "savings_goal": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "entities.ResponseModel": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}